requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
